# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Parking POC API
  version: 0.1.0
  x-contract-status: frozen  # M0-frozen — changes only via reviewed PR
  description: >
    Office parking-slot booking platform. Contract-first source of truth (frozen at M0 — changes via
    reviewed PR). Conventions, envelope, and error codes per docs/phases-3-6-plan.md §3.1–3.2.
    Maturity of each operation is marked with `x-maturity` (DEMO | MVP | Later).
servers:
  - url: /api/v1
    description: Base path (path-based versioning)

# Default: every operation requires a bearer access token. Public endpoints override with `security: []`.
security:
  - bearerAuth: []
...
paths:
...
  /bookings:
    get:
      tags: [Bookings]
      operationId: listBookings
      summary: List bookings for admins (paged, newest first)
      description: >
        Role COMPANY_ADMIN (scoped to their own company) or SUPER_ADMIN (all companies; optional
        `companyId` filter). Each row carries who booked, the date, status, and allocated slot.
        Filter by `date` and/or `status`.
      x-maturity: DEMO
      x-required-roles: [COMPANY_ADMIN, SUPER_ADMIN]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: date
          in: query
          required: false
          description: Filter to a single booking date (YYYY-MM-DD).
          schema: { type: string, format: date }
        - name: companyId
          in: query
          required: false
          description: SUPER_ADMIN only — narrow to one company. Ignored for COMPANY_ADMIN.
          schema: { type: string }
        - name: status
          in: query
          required: false
          description: Optional filter by booking status.
          schema: { $ref: '#/components/schemas/BookingStatus' }
      responses:
        '200':
          description: Paged list of bookings.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [meta]
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AdminBooking' }
                      meta: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Bookings]
      operationId: createBooking
      summary: Submit a parking request for the next bookable weekday
      description: >
        Role: USER. Creates a PRIMARY booking for a bookable weekday (Mon–Fri, D7). `travelDistanceKm` is
        snapshotted from the user's profile at submission (F6); `carpoolPeople` includes the driver as person 1
        (D3) and the server enforces the current `carpool.maxPeople` cap (default 4, D8). Only same-company
        employee carpool members are scored (F4). Errors: 422 WINDOW_CLOSED if past `booking.primaryCutoff` or
        a non-bookable date; 409 CONFLICT on a duplicate same-type request for the date.
      x-maturity: DEMO
      x-required-roles: [USER]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingRequest' }
            example:
              bookingDate: "2026-07-31"
              vehicleType: CAR
              vehicleNumber: MH12AB1234
              carpoolPeople: 3
              specialRequirement: null
              carpoolMembers:
                - { name: Rahul Mehta, employeeEmail: rahul@assent.example }
                - { name: Neighbour, employeeEmail: null }
      responses:
        '201':
          description: Booking submitted.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BookingCreatedData' }
              example:
                success: true
                data:
                  id: bkg_123
                  status: SUBMITTED
                  bookingType: PRIMARY
                  bookingDate: "2026-07-31"
                  travelDistanceKm: 12.4
                  carpoolPeople: 3
                  submittedAt: "2026-07-30T11:20:00Z"
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/WindowClosed' }
...
  /bookings/{id}:
    parameters:
      - $ref: '#/components/parameters/BookingIdParam'
    get:
      tags: [Bookings]
      operationId: getBooking
      summary: Get a booking's status, allocated slot, and score breakdown
      description: "Role: USER (own) / COMPANY_ADMIN (own company) / SUPER_ADMIN. Returns 404 if not visible to the tenant."
      x-maturity: DEMO
      x-required-roles: [USER, COMPANY_ADMIN, SUPER_ADMIN]
      responses:
        '200':
          description: Booking detail incl. allocated slot and per-factor score breakdown (once allocation has run).
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BookingDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Bookings]
      operationId: updateBooking
      summary: Edit a booking before the primary cutoff
      description: "Role: USER (own). Partial edit; allowed only while `now < booking.primaryCutoff` (else 422 WINDOW_CLOSED)."
      x-maturity: MVP
      x-required-roles: [USER]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBookingRequest' }
      responses:
        '200':
          description: Booking updated.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BookingDetail' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/WindowClosed' }
...
  /bookings/{id}/release:
    parameters:
      - $ref: '#/components/parameters/BookingIdParam'
    post:
      tags: [Bookings]
      operationId: releaseBooking
      summary: Release an allocated slot
      description: >
        Role: USER (own) or COMPANY_ADMIN (own users). Releasing an allocated slot after primary allocation
        triggers reallocation to the highest-scoring own-company waitlisted user first, else the slot moves to
        the common pool (F3). Audited + notified.
      x-maturity: MVP
      x-required-roles: [USER, COMPANY_ADMIN]
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReleaseBookingRequest' }
      responses:
        '200':
          description: Slot released (reallocation handled server-side, F3).
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BookingDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  # ---------------------------------------------------------------------------
  # Allocation  (P3-07)  — Role: SUPER_ADMIN.
  # ---------------------------------------------------------------------------
...
  /availability:
    get:
      tags: [Bookings]
      operationId: getAvailability
      summary: Per-date slot grid for the caller's company + booking-window summary
      description: >
        Any authenticated role. Returns one entry per date in `[from, to]` (defaulting to exactly the
        currently-open window) with the company's quota, how much is blocked/taken, and a `boxes`
        array to render as the slot grid. The company always comes from the principal, never the
        query. `taken` counts LIVE REQUESTS, not just allocations (D12) — a submitted request holds a
        box immediately, which is what keeps demand within supply so nobody is rejected later.
      x-maturity: MVP
      x-required-roles: [USER, COMPANY_ADMIN, SUPER_ADMIN, SECURITY]
      parameters:
        - name: from
          in: query
          required: false
          description: First date (YYYY-MM-DD). Defaults to the first requestable date.
          schema: { type: string, format: date }
        - name: to
          in: query
          required: false
          description: Last date (YYYY-MM-DD). Defaults to the end of the window. Max span 120 days.
          schema: { type: string, format: date }
      responses:
        '200':
          description: Window summary plus per-date availability.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AvailabilityResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from `POST /auth/login`, sent as `Authorization: Bearer <token>`."
  headers:
    CorrelationId:
      description: Request/response correlation id (echoed into AuditLog).
      schema: { type: string }
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: Items per page (max 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    BookingIdParam:
      name: id
      in: path
      required: true
      description: Booking request id.
      schema: { type: string }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Standard success envelope. `data` is the resource/list; `meta` present only on list endpoints.
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        # No `type` here: each endpoint overrides `data` (object or array) via allOf.
        data:
          description: Resource or list payload; concrete shape defined per endpoint.
        meta: { $ref: '#/components/schemas/Pagination' }
    ErrorEnvelope:
      type: object
      description: Standard error envelope.
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { $ref: '#/components/schemas/ErrorObject' }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ErrorCode:
      type: string
      description: Canonical error codes (docs/phases-3-6-plan.md §3.2).
      enum:
        - VALIDATION_ERROR   # 400
        - UNAUTHENTICATED    # 401
        - FORBIDDEN          # 403
        - NOT_FOUND          # 404
        - CONFLICT           # 409
        - CAPACITY_FULL      # 409 — every slot for the date is held (Phase 7 D12)
        - WINDOW_CLOSED      # 422
        - RATE_LIMITED       # 429
        - INTERNAL           # 500
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }

    # --- Enums (from prisma/schema.prisma) ---
...
    BookingType:
      type: string
      enum: [PRIMARY, COMMON_POOL]
    BookingStatus:
      type: string
      enum: [DRAFT, SUBMITTED, CANCELLED, ALLOCATED, WAITLISTED, REJECTED, RELEASED, EXPIRED]
    VehicleType:
      type: string
      enum: [CAR, BIKE, EV_CAR, EV_BIKE, OTHER]
...
    Booking:
      type: object
      required: [id, bookingDate, bookingType, status, carpoolMemberCount, createdAt]
      properties:
        id: { type: string }
        bookingDate: { type: string, format: date }
        bookingType: { $ref: '#/components/schemas/BookingType' }
        status: { $ref: '#/components/schemas/BookingStatus' }
        travelDistanceKm: { type: number, nullable: true }
        vehicleType:
          $ref: '#/components/schemas/VehicleType'
        vehicleNumber: { type: string, nullable: true }
        carpoolMemberCount: { type: integer, minimum: 0 }
        specialRequirement: { type: string, nullable: true }
        allocationScore: { type: number, nullable: true }
        allocatedSlotNumber:
          type: string
          nullable: true
          description: Slot number if ALLOCATED, else null.
        submittedAt: { type: string, nullable: true, format: date-time }
        createdAt: { type: string, format: date-time }
    AdminBooking:
      type: object
      description: A booking row for the admin dashboards — who booked, when, its status and slot.
      required: [id, bookingDate, bookingType, allocationSource, status, employeeName, employeeEmail, companyId, companyName, carpoolPeople, createdAt]
      properties:
        id: { type: string }
        bookingDate: { type: string, format: date }
        bookingType: { $ref: '#/components/schemas/BookingType' }
        allocationSource:
          type: string
          nullable: true
          enum: [PRIMARY, COMMON_POOL, RELEASED_SLOT, MANUAL_OVERRIDE]
          description: Actual source of the allocated seat; null when no seat is assigned.
        status: { $ref: '#/components/schemas/BookingStatus' }
        employeeName: { type: string }
        employeeEmail: { type: string, format: email }
        companyId: { type: string }
        companyName: { type: string }
        travelDistanceKm: { type: number, nullable: true }
        carpoolPeople: { type: integer, minimum: 1, description: Driver + declared members. }
        carpoolMembers:
          type: array
          description: Declared passengers carried by the driver.
          items:
            type: object
            required: [id, name, sameCompany, isScored]
            properties:
              id: { type: string }
              name: { type: string }
              employeeEmail: { type: string, format: email, nullable: true }
              contactNumber: { type: string, nullable: true }
              pickupLocation: { type: string, nullable: true }
              sameCompany: { type: boolean }
              isScored: { type: boolean }
        allocationScore: { type: number, nullable: true }
        allocatedSlotNumber: { type: string, nullable: true, description: Slot number if ALLOCATED, else null. }
        submittedAt: { type: string, nullable: true, format: date-time }
        createdAt: { type: string, format: date-time }
        history:
          type: array
          description: Same employee/date booking lifecycle rows, grouped before pagination.
          items: { $ref: '#/components/schemas/AdminBooking' }
...
    BookingWindow:
      type: object
      description: >
        The rolling request window and the weekly run it closes at (Phase 7 §3). Everything the client
        needs to say "book between X and Y; results published on Z".
      required:
        [nextRunAt, nextRunCountdownSeconds, requestCloseAt, requestCloseCountdownSeconds, resultsRunAt,
         runDay, runFrequency, runTime, windowWeeks, approvalLeadDays, earliestDate, latestDate,
         requestableDates, nextRuns]
      properties:
        nextRunAt: { type: string, format: date-time, description: When the next automatic batch runs. }
        nextRunCountdownSeconds: { type: integer, description: Seconds until `nextRunAt`; 0 once passed. }
        requestCloseAt: { type: string, format: date-time, description: When requests close for the next processed band. }
        requestCloseCountdownSeconds: { type: integer, description: Seconds until `requestCloseAt`; 0 once passed. }
        resultsRunAt: { type: string, format: date-time, description: Run that publishes results for the open request band. }
        runDay: { type: string, enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY] }
        runFrequency: { type: string, enum: [WEEKLY, BIWEEKLY, MONTHLY] }
        runTime: { type: string, description: 'HH:MM IST.' }
        windowWeeks: { type: integer, enum: [2, 4] }
        approvalLeadDays: { type: integer, description: Days a date is decided ahead of itself (D11). }
        earliestDate: { type: string, format: date }
        latestDate: { type: string, format: date }
        requestableDates:
          type: array
          description: Bookable weekdays currently open for requests.
          items: { type: string, format: date }
        nextRuns:
          type: array
          description: Next five automatic allocation run instants.
          items: { type: string, format: date-time }
    SlotBoxState:
      type: string
      description: >
        MINE = the caller's own reservation, TAKEN = someone else's, BLOCKED = withheld by an admin,
        AVAILABLE = free. Render AVAILABLE green and everything else grey.
      enum: [AVAILABLE, TAKEN, BLOCKED, MINE]
    SlotBox:
      type: object
      required: [index, state]
      properties:
        index: { type: integer, description: 1-based position in the grid. }
        state: { $ref: '#/components/schemas/SlotBoxState' }
    UnavailableReason:
      type: string
      description: Why a date cannot be requested; `null` when it can.
      enum: [INVALID_DATE, NOT_WEEKDAY, TOO_SOON, BEYOND_WINDOW, NO_QUOTA, FULL, ALREADY_BOOKED]
    DayAvailability:
      type: object
      required: [date, quota, blocked, taken, available, mine, requestable, boxes]
      properties:
        date: { type: string, format: date }
        quota:
          type: integer
          description: The company's effective quota for this date — also the number of boxes (D14).
        blocked: { type: integer }
        taken:
          type: integer
          description: Live requests holding a box (D12), not merely completed allocations.
        available: { type: integer }
        mine: { type: boolean, description: True when the caller already holds a request for this date. }
        requestable: { type: boolean }
        reason:
          allOf: [ { $ref: '#/components/schemas/UnavailableReason' } ]
          nullable: true
        message: { type: string, nullable: true, description: Human-readable explanation of `reason`. }
        boxes:
          type: array
          items: { $ref: '#/components/schemas/SlotBox' }
    AvailabilityResponse:
      type: object
      required: [window, days]
      properties:
        window: { $ref: '#/components/schemas/BookingWindow' }
        days:
          type: array
          items: { $ref: '#/components/schemas/DayAvailability' }
...
    CarpoolMemberInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        employeeEmail:
          type: string
          format: email
          nullable: true
          description: Same-company employee email; only validated same-company employees are scored (F4).
        contactNumber: { type: string, nullable: true }
        pickupLocation: { type: string, nullable: true }
    CarpoolMember:
      type: object
      required: [name, sameCompany, isScored]
      properties:
        id: { type: string }
        name: { type: string }
        employeeEmail: { type: string, format: email, nullable: true }
        sameCompany: { type: boolean }
        isScored: { type: boolean, description: True only for validated same-company employees (F4). }
    CreateBookingRequest:
      type: object
      required: [bookingDate, carpoolPeople]
      properties:
        bookingDate: { type: string, format: date, description: "Target bookable weekday (Mon–Fri, D7)." }
        vehicleType: { $ref: '#/components/schemas/VehicleType' }
        vehicleNumber: { type: string, nullable: true }
        carpoolPeople:
          type: integer
          minimum: 1
          description: >
            Total people incl. the driver (person 1, D3). Server enforces the current carpool.maxPeople cap
            (default 4, D8) — no fixed maximum is baked into the contract so a config change needs no re-spec.
        specialRequirement: { type: string, nullable: true }
        carpoolMembers:
          type: array
          description: Optional carpool members; only same-company employees are scored (F4).
          items: { $ref: '#/components/schemas/CarpoolMemberInput' }
    UpdateBookingRequest:
      type: object
      description: Partial edit before the primary cutoff. All fields optional.
      properties:
        vehicleType: { $ref: '#/components/schemas/VehicleType' }
        vehicleNumber: { type: string, nullable: true }
        carpoolPeople: { type: integer, minimum: 1 }
        specialRequirement: { type: string, nullable: true }
        carpoolMembers:
          type: array
          items: { $ref: '#/components/schemas/CarpoolMemberInput' }
...
    ReleaseBookingRequest:
      type: object
      properties:
        reason: { type: string, nullable: true }
    BookingCreatedData:
      type: object
      required: [id, status, bookingType, bookingDate, carpoolPeople]
      properties:
        id: { type: string }
        status: { $ref: '#/components/schemas/BookingStatus' }
        bookingType: { $ref: '#/components/schemas/BookingType' }
        bookingDate: { type: string, format: date }
        travelDistanceKm:
          type: number
          nullable: true
          description: Snapshotted from the user's profile at submission (F6).
        carpoolPeople: { type: integer }
        submittedAt: { type: string, format: date-time, nullable: true }
    ScoreBreakdown:
      type: object
      description: Per-factor allocation scoring (decisions §3). Present once an allocation run has scored the booking.
      required: [people, distanceScore, carpoolScore, distanceWeight, carpoolWeight, finalScore]
      properties:
        distanceKm: { type: number, nullable: true, description: Snapshotted travel distance used for scoring. }
        people: { type: integer, description: "1 (driver) + scored same-company carpool members (F4)." }
        distanceScore: { type: number, description: "(min(distanceKm, maxDistanceKm)/maxDistanceKm)*100." }
        carpoolScore: { type: number, description: "((min(people, maxPeople)-1)/(maxPeople-1))*100." }
        distanceWeight: { type: number }
        carpoolWeight: { type: number }
        finalScore: { type: number, description: "distanceWeight*distanceScore + carpoolWeight*carpoolScore." }
    BookingDetail:
      allOf:
        - $ref: '#/components/schemas/Booking'
        - type: object
          properties:
            carpoolMembers:
              type: array
              items: { $ref: '#/components/schemas/CarpoolMember' }
            scoreBreakdown:
              allOf: [ { $ref: '#/components/schemas/ScoreBreakdown' } ]
              description: Per-factor scores; present only after an allocation run has scored this booking (absent/null before).

    # --- Allocation (P3-07) ---
...
  responses:
...
    ValidationError:
      description: 400 — request validation failed.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details: [ { field: distanceKm, message: must be between 0 and 200 } ]
    Unauthorized:
      description: 401 — missing/invalid/expired access token or bad credentials.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: UNAUTHENTICATED, message: Invalid credentials } }
    Forbidden:
      description: 403 — role or tenant-scope violation (or inactive account).
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: FORBIDDEN, message: Account is not active } }
    NotFound:
      description: 404 — resource absent or not visible to the tenant.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: NOT_FOUND, message: Resource not found } }
    Conflict:
      description: 409 — duplicate resource.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: CONFLICT, message: Email already registered } }
...
    WindowClosed:
      description: 422 — action attempted outside its allowed time window (e.g. after `booking.primaryCutoff`, or a non-bookable date).
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: WINDOW_CLOSED, message: The booking window is closed for this date } }